Stock fetching converts to csv and then json. Functions correctly.
