+++ src/InvestmentSelect/InvestmentSelect.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import Form from 'react-bootstrap/Form';
 import styles from './InvestmentSelect.module.css';
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
 
 const InvestmentSelect = ({onInvestmentSelect, onReinvestIntervalSelect}) => {
     function handleInvestmentSelect(event) {
