--- src/data/fetchStockJsonApi.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-async function fetchAPI(ticker) {
-    return await axios.get('http://127.0.0.1:5000/stock/' + ticker)
-}
-
-
-export default fetchAPI;
